src/requirements.txt
google-generativeai
python-dotenv
numpy
pandas
scikit-learn
flask
requests